refactor: searchQuery 관련 비즈니스 로직을 분리
